// src/app/shared/types/api-types/bire-concerned-asset-dto.interface.ts
export interface BireConcernedAssetDTO {
  assetId?: string;
  evolutionNumber?: string;
  evolutionRevisionNumber?: string;
  sbNumber?: string;
  sbRevisionNumber?: string;
  familyCode?: string;
  structureType?: string;
  modificationType?: string;
  modificationNumber?: string;
  modificationRevisionNumber?: string;
  assetPn?: string;
  assetSn?: string;
  assetActionType?: string;
}